[SPARK-40622][SQL][CORE] Remove the limitation that single task result must fit in 2GB
### What changes were proposed in this pull request?

Single task result must fit in 2GB, because we're using byte array or `ByteBuffer`(which is backed by byte array as well), and thus has a limit of 2GB(java array size limit on `byte[]`).

This PR is trying to fix this by replacing byte array with `ChunkedByteBuffer`.

### Why are the changes needed?

To overcome the 2GB limit for single task.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Unit test

Closes #38064 from liuzqt/SPARK-40622.

Authored-by: Ziqi Liu <ziqi.liu@databricks.com>
Signed-off-by: Mridul <mridul<at>gmail.com>