[SPARK-41189][PYTHON] Add an environment to switch on and off namedtuple hack
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/34688 that adds a switch to turn on and off the namedtuple hack.

### Why are the changes needed?

There are still behaviour differences between regular pickle and Cloudpickle e.g., bug fixes from the upstream. It's safer to have a switch to turn on and off for the time being.

### Does this PR introduce _any_ user-facing change?

This remains as an internal environment so ideally no. In fact the main change itself was the internal change too.

### How was this patch tested?

Manually tested.

Closes #38700 from HyukjinKwon/SPARK-41189.

Lead-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>