[SPARK-41224][SPARK-41165][SPARK-41184] Optimized Arrow-based collect implementation to stream from server to client
### What changes were proposed in this pull request?

This PR proposes an optimized Arrow-based collect, that is virtually https://github.com/apache/spark/pull/38720 that implements the logics except a couple of nits.

### Why are the changes needed?

To stream the Arrow batch from the server to the client side instead of waiting all the jobs to finish.

### Does this PR introduce _any_ user-facing change?

No, this feature isn't released yet.

### How was this patch tested?

Unittest added.

Closes #38720

Closes #38759 from HyukjinKwon/SPARK-41165-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Herman van Hovell <herman@databricks.com>