[SPARK-41297][CONNECT][PYTHON] Support String Expressions in filter
### What changes were proposed in this pull request?
This patch extends the functionality of the DataFrame in Spark Connect by support passing string based SQL expressions to the filter. The expressions are passed as is to the server during evaluation. If the expression is not boolean an Analysis error is thrown.

```
spark.range(100).filter("id < 10").collect()
```

### Why are the changes needed?
Compatibility

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #38826 from grundprinzip/SPARK-41297.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>