[SPARK-41325][CONNECT] Fix missing avg() for GroupBy on DF
### What changes were proposed in this pull request?
Previously, the `avg` function was missing in the `GroupedData` class. This patch adds this method and the necessary plan transformation using an unresolved function. In addition, it identified a small issue where when an alias is used for a grouping column, the planner would incorrectly try to wrap the existing alias expression using an unresolved alias which would then fail.

```
df = (
    self.connect.range(10)
    .groupBy((col("id") % lit(2)).alias("moded"))
    .avg("id")
    .sort("moded")
)
```

### Why are the changes needed?
Bug / Compatibility

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #38841 from grundprinzip/SPARK-41325.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>