[SPARK-40379][K8S] Propagate decommission executor loss reason in K8s
Propagate decommission executor loss reason in K8s during onDisconnect

Currently if an executor has been sent a decommission message and then it disconnects from the scheduler we only disable the executor depending on the K8s status events to drive the rest of the state transitions. However, the K8s status events can become overwhelmed on large clusters so we should check if an executor is in a decommissioning state when it is disconnected and use that reason instead of waiting on the K8s status events so we have more accurate logging information.

Logging output will change.

Existing unit tests.

Closes #37821 from holdenk/SPARK-40379-progegate-decommission-exec-loss-reason.

Authored-by: Holden Karau <holden@pigscanfly.ca>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>