[SPARK-41402][SQL][CONNECT] Override prettyName of StringDecode
### What changes were proposed in this pull request?

This PR proposes to change `prettyName` of `StringDecode` to `decode` to keep the name as was. It was changed from SPARK-33527.

Note that I don't change `nodeName` here to keep the strings in the plan to differentiate `Decode` from `StringDecode`.

### Why are the changes needed?

This is technically a beahviour change (or regression) from SPARK-33527 although we don't guarantee the compatibility there.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the default output column name from `stringdecode` to `decode`.

### How was this patch tested?

Relates unittests were fixed together.

Closes #39001 from HyukjinKwon/SPARK-41402.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>