[SPARK-41463][SQL][TESTS] Ensure error class names contain only capital letters, numbers and underscores
### What changes were proposed in this pull request?
This change adds a unit test to verify that error class and subclass names contain only capital letters, numbers and underscores.

### Why are the changes needed?
This is to ensure that the standard holds for error classes added in the future.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Test only change. Also tried adding an invalid error class name and verified that the test failed locally.

Closes #38998 from bozhang2820/spark-41463.

Authored-by: Bo Zhang <bo.zhang@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>